build(release): Ignore unsupported platform architecture combinations
fix(tuning): Fix type conversion issue in memory calculation

+++ internal/shared/tuning/mem_linux.go
@@ -7,7 +7,7 @@ import "syscall"
 func getSystemTotalMemory() uint64 {
 	var info syscall.Sysinfo_t
 	if err := syscall.Sysinfo(&info); err == nil {
-		return info.Totalram * uint64(info.Unit)
+		return uint64(info.Totalram) * uint64(info.Unit)
 	}
 	return 1024 * 1024 * 1024
 }
